+++ lib/ts/recipe/emailpassword/components/themes/translations.ts
@@ -82,5 +82,9 @@ export const defaultTranslationsEmailPassword = {
         "Email is invalid": undefined,
         "Reset password link was not created because of account take over risk. Please contact support. (ERR_CODE_001)":
             undefined,
+        "Cannot sign up due to security reasons. Please try logging in, use a different login method or contact support. (ERR_CODE_007)":
+            undefined,
+        "Cannot sign in due to security reasons. Please try resetting your password, use a different login method or contact support. (ERR_CODE_008)":
+            undefined,
     },
 };